Consolidate samples into a single package.
